-moved Extendable in the types/utils -updated all components with correct imports -added typings for the AccordionTitle and AccordionContent
