+++ src/components/Accordion/AccordionContent.tsx
@@ -2,11 +2,21 @@ import * as PropTypes from 'prop-types'
 import * as React from 'react'
 
 import { childrenExist, createShorthandFactory, customPropTypes, UIComponent } from '../../lib'
+import { Extendable } from '../../../types/utils'
+
+export interface IAccordionContentProps {
+  as?: any
+  active?: boolean
+  children?: React.ReactNode
+  className?: string
+  content?: React.ReactNode
+  onClick?: (event: React.SyntheticEvent, data: IAccordionContentProps) => void
+}
 
 /**
  * A standard AccordionContent.
  */
-class AccordionContent extends UIComponent<any, any> {
+class AccordionContent extends UIComponent<Extendable<IAccordionContentProps>, any> {
   static displayName = 'AccordionContent'
 
   static create: Function
