+++ src/components/Grid/Grid.tsx
@@ -1,15 +1,9 @@
 import * as PropTypes from 'prop-types'
 import * as React from 'react'
 import ReactNode = React.ReactNode
-import {
-  UIComponent,
-  childrenExist,
-  customPropTypes,
-  IRenderResultConfig,
-  Extendable,
-} from '../../lib'
+import { UIComponent, childrenExist, customPropTypes, IRenderResultConfig } from '../../lib'
 import { ComponentVariablesInput, IComponentPartStylesInput } from '../../../types/theme'
-import { ItemShorthand } from '../../../types/utils'
+import { Extendable, ItemShorthand } from '../../../types/utils'
 
 export interface IGridProps {
   as?: any
