--- src/lib/Extendable.ts
@@ -1,3 +1,0 @@
-export type Extendable<T> = T & {
-  [key: string]: any
-}
