+++ src/lib/index.ts
@@ -39,4 +39,3 @@ export { default as leven } from './leven'
 
 export { pxToRem, setHTMLFontSize } from './fontSizeUtility'
 export { customPropTypes, SUI }
-export { Extendable } from './Extendable'
