+++ types/utils.d.ts
@@ -1,4 +1,12 @@
 // ========================================================
+// Utilities
+// ========================================================
+
+export type Extendable<T> = T & {
+  [key: string]: any
+}
+
+// ========================================================
 // Props
 // ========================================================
